refactor(client): generator
