--- client/src/app/components/fractal/fractal.css
@@ -1,50 +1,0 @@
-:host {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 1rem;
-  justify-content: center;
-}
-
-.controls {
-  display: flex;
-  flex-direction: column;
-  row-gap: 0.5rem;
-}
-
-.btns {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 0.5rem;
-}
-
-canvas {
-  border: var(--border-thickness) solid var(--color40);
-  border-radius: var(--border-radius);
-  box-shadow: 0 0 15px rgba(109, 213, 250, 0.4);
-}
-
-.control_btn,
-.color_btn {
-  width: 2rem;
-  height: 2rem;
-  border-radius: var(--border-radius);
-  cursor: pointer;
-  transition: transform 0.25s linear, box-shadow 0.25s linear;
-  box-shadow: 2px 2px 3px #888;
-
-  &:hover {
-    transform: translate(1px, 1px);
-    box-shadow: 1px 1px 2px #888;
-  }
-}
-
-.control_btn {
-  background-color: #222;
-  filter: invert(var(--color-scheme));
-}
-
-.color_btn {
-  --from: #000;
-  --to: #fff;
-  background: linear-gradient(var(--from), var(--to));
-}
